requirements.txt
# Core ML and NLP libraries
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1
faiss-cpu>=1.7.2
sentence-transformers>=2.2.2
torch>=2.0.0
numpy>=1.23.0

# PDF generation and processing
fpdf2>=2.7.0
PyMuPDF>=1.23.0

# Progress tracking and utilities
tqdm>=4.64.0
pandas>=1.5.0

# Note: Built-in Python modules (textwrap, re, string, pathlib, typing, json, os, math) 
# are included with Python and don't need to be installed separately
